added more pages, added randomizatio of cards on each game and other inprovements
