+++ .gitignore
@@ -31,3 +31,4 @@ typings/
 !.vscode/tasks.json
 !.vscode/launch.json
 !.vscode/extensions.json
+launch.json
