+++ package.json
@@ -28,7 +28,9 @@
     "zone.js": "~0.9.1"
   },
   "devDependencies": {
+    "@angular/cli": "^8.3.8",
     "@angular/compiler-cli": "~8.2.0",
+    "@nativescript/schematics": "^0.7.2",
     "@ngtools/webpack": "~8.2.0",
     "nativescript-dev-webpack": "~1.2.0",
     "typescript": "~3.5.3"
