+++ src/app/app.module.ts
@@ -4,6 +4,8 @@ import { NativeScriptModule } from "nativescript-angular/nativescript.module";
 import { AppRoutingModule } from "./app-routing.module";
 import { AppComponent } from "./app.component";
 import { HomeComponent } from "./home/home.component";
+import { GameSelectionComponent } from './game-selection/game-selection.component';
+import { GameBoardComponent } from './game-board/game-board.component';
 
 // Uncomment and add to NgModule imports if you need to use two-way binding
 // import { NativeScriptFormsModule } from "nativescript-angular/forms";
@@ -21,7 +23,9 @@ import { HomeComponent } from "./home/home.component";
     ],
     declarations: [
         AppComponent,
-        HomeComponent
+        HomeComponent,
+        GameSelectionComponent,
+        GameBoardComponent
     ],
     providers: [],
     schemas: [
