+++ src/app/game-board/game-board.component.html
@@ -1,0 +1,11 @@
+<GridLayout class="page" columns="*,*,*,*" rows="*,*,*,*">
+  <Button
+    *ngFor="let tile of tiles"
+    (tap)="onTapHandler(tile)"
+    [col]="tile.col"
+    [row]="tile.row"
+    [backgroundColor]="tile.shownColor"
+    [text]="tile.text"
+  ></Button>
+  <!-- Add your page content here -->
+</GridLayout>
