+++ src/app/game-selection/game-selection.component.css
@@ -1,0 +1,1 @@
+/* Add mobile styles for the component here.  */
